+++ src/main/groovy/com/cedarsoftware/ncube/NCubeJdbcPersisterAdapter.groovy
@@ -250,6 +250,12 @@ class NCubeJdbcPersisterAdapter implements NCubePersister
                 "search(${appId}, ${cubeNamePattern}, ${searchValue})")
     }
 
+    List<NCube> getPullRequestCubes(ApplicationID appId, Date startDate, Date endDate)
+    {
+        return (List<NCube>) jdbcOperation({ Connection c -> persister.getPullRequestCubes(c, appId, startDate, endDate) },
+                "getPullRequestCubes(${appId}, ${startDate}, ${endDate})")
+    }
+
     void clearTestDatabase()
     {
         jdbcOperation({ Connection c -> persister.clearTestDatabase(c) },